BusinessRules eklendi. Kredi Kartı için gereksinimler sağlandı.

+++ Business/Concrete/CarImageManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -7,6 +8,7 @@ using Business.Abstract;
 using Business.Constants;
 using Business.ValidationRules.FluentValidation;
 using Core.Aspects.Autofac.Validation;
+using Core.Utilities.Business;
 using Core.Utilities.Helpers;
 using Core.Utilities.Results;
 using DataAccess.Abstract;
@@ -42,6 +44,13 @@ namespace Business.Concrete
         [ValidationAspect(typeof(CarImageValidator))]
         public IResult Add(IFormFile file, CarImage carImage)
         {
+            IResult result = BusinessRules.Run(CheckIfImageLimitExceeded(carImage.CarId),CheckIfImageExtensionValid(file));
+
+            if (result !=null)
+            {
+                return result;
+            }
+
             carImage.ImagePath = FileHelper.Add(file);
             carImage.Date = DateTime.Now;
             _carImageDal.Add(carImage);
@@ -51,14 +60,29 @@ namespace Business.Concrete
         [ValidationAspect(typeof(CarImageValidator))]
         public IResult Update(IFormFile file, CarImage carImage)
         {
+            IResult result = BusinessRules.Run(CheckIfImageExists(carImage.Id), CheckIfImageExtensionValid(file));
+
+            if (result != null)
+            {
+                return result;
+            }
+
             carImage.ImagePath = FileHelper.Update(_carImageDal.Get(p=>p.Id==carImage.Id).ImagePath,file);
             carImage.Date = DateTime.Now;
             _carImageDal.Update(carImage);
             return new SuccessResult(Messages.CarImageUpdated);
         }
-
+        
+        [ValidationAspect(typeof(CarImageValidator))]
         public IResult Delete(CarImage carImage)
         {
+            IResult result = BusinessRules.Run(CheckIfImageExists(carImage.Id));
+
+            if (result != null)
+            {
+                return result;
+            }
+
             FileHelper.Delete(carImage.ImagePath);
             _carImageDal.Delete(carImage);
             return new SuccessResult(Messages.CarImageDeleted);
@@ -87,6 +111,28 @@ namespace Business.Concrete
             }
 
             return _carImageDal.GetList(p => p.CarId == id);
+        }
+
+        private IResult CheckIfImageExtensionValid(IFormFile file)
+        {
+            bool isValidFileExtension =
+                Messages.CarImageValidationFileTypes.Any(t => t == Path.GetExtension(file.FileName).ToUpper());
+            if (!isValidFileExtension)
+            {
+                return new ErrorResult(Messages.CarImageExtensionInvalid);
+            }
+
+            return new SuccessResult();
+        }
+
+        private IResult CheckIfImageExists(int id)
+        {
+            if (_carImageDal.IsExists(id))
+            {
+                return new SuccessResult();
+            }
+
+            return new ErrorResult(Messages.CarImageNotExists);
         }
     }
 }
